pyproject.toml
[project]
name = "planning-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "langchain-core>=0.2.38",
    "langchain-community>=0.2.16",
    "langchain-openai>=0.1.23",
    "pillow>=10.4.0",
    "polars>=1.6.0",
    "fastexcel>=0.11.6",
    "spacy>=3.7.6",
    "pip>=24.2",
    "spacytextblob>=4.0.0",
    "transformers>=4.44.2",
    "langgraph>=0.2.18",
    "pdf2image>=1.17.0",
]

[tool.uv]
dev-dependencies = [
    "ipdb>=0.13.13",
    "ipython>=8.27.0",
]
